backend/requirements.txt
# Web framework
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pgvector

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.20

# Environment & config
pydantic-settings==2.7.0
python-dotenv==1.0.1

# Pydantic
pydantic[email]==2.10.4

# Supabase Storage
supabase==2.10.0

# File handling & validation
filetype==1.2.0

# RAG & AI Pipeline
llama-index-core
llama-index-readers-file
llama-index-embeddings-huggingface
sentence-transformers
groq
openai

# HTTP
httpx